+++ backend_flask/app/api/__init__.py
@@ -1,0 +1,6 @@
+from flask import Blueprint
+
+api = Blueprint('api', __name__)
+
+from . import authentication, books, decorators, errors, users
+
