+++ backend_flask/app/api/authentication.py
@@ -1,0 +1,19 @@
+from flask import g
+
+from flask_httpauth import HTTPBasicAuth
+from ..models import User
+from . import api
+
+
+auth = HTTPBasicAuth()
+
+@auth.verify_password
+def verify_password(email, password):
+    if email == '':
+        return False
+    user = User.query.filter_by(email = email).first()
+    if not user:
+        return False
+    g.current_user = user
+    return user.verify_password(password)
+
